Tucking away these changes which inline more constructor stuff
[this change is highly speculative]

This should be double-checked before committing it, but I think
what I've done is to inline the setup of ranges and of other
default-constructed records.  My thinking was to give the C
compiler more information (and also make it simpler for me to
read in the generated code what was happening when ranges are
created.