Renaming this .txt file to .md
While it may have sounded like I was anti-markdown in our meeting the
other week, an important caveat is that it was only in the context of
documents that would most likely be viewed by the user in an
environment that didn't support markdown where markdown made things
uglier.  This file doesn't meet that criteria, being created for
developers and therefore likely to be viewed on github (besides which,
given that it's already been converted to markdown, and with
hyperlinks, it's ugly as a .txt file).