Removing inline-oriented suppressions from baseline.suppress
For some time now, we've thrown --inline when doing --baseline
testing for reasons that I can't quite recall.  This has caused
tests that used to be suppressed for baseline testing (because
they test that inlining is working) to result in errors because
the expected suppression wasn't found.

This commit comments out those suppressions and adds an
explanation such that, if/when we make --baseline testing
just throw --baseline again, the history is visible in the
file.