Make Random generic with respect to an eltType
In documenting RandomStream, it showed indications of its evolution.
Specifically, its routines for getting the next random value were very
real(64) based, while fillRandom() could take any of real(64),
imag(64), or complex(128).  A better approach seemed to be to
parameterize the stream by an eltType, defaulting to real(64) for
backwards compatibility.  This also permitted some cases to be cleaned
up in RandomStream.fillRandom() by assuming that it could only take an
array of the given element type.