This..... doesn't quite work...
But it compiles!  The thinking here is that I don't
need to create a new Block distribution (which was
the path that this afternoon's attempt started down),
simply a meta-descriptor that knows how to create a
block array over a given domain and to reindex to
it just as we do for the original array.  This seems
simpler.

The main issue seems to relate to privatized classes
and accessing the _value fields directly, which is
probably not kosher on a privatized class.  Could
take more care to either access the field or not, but...
what if we simply stored a record in the dom/arr
fields instead and used their official interface.
Wouldn't that be easier?  Although... it would mean
implementing the dsi interface using the non-dsi
interface.

Suggested next step: Turn privatization off and see
if I can get that working; then turn it on and see
if I'm more careful with the _values whether I can get
that working...