Stopping point for tonight
Got things in the domain class a bit more principled.

Am running into strage issues in the reindex views...  For
a --no-local compilation, someone is trying to get a privatized
class with ID of -1.  For the 2-locale GASNet case, I'm getting
a warning about referring to remote data in a local block.