Making the domain class serve as both domain and distribution
It turns out that the trouble of my woes was:

  var b: a.type;

where a was one of my array view types.  Because I had these
domains forward their 'dist' query on to their base domain,
the type of 'b' was made to be a domain of that type.

Because the role of a distribution is to create new domains
of that type, I took the approach of making the domain class
also serve as a distribution and to know how to create things
like it.

This required adding dsiSetIndices() to the domain as well
as the reprivatize interface since it will be reprivatized,
obviously.

Now I'm getting segfaults, but I'm pretty sure I didn't get
reprivatize quite right...  In particular, I ignored 'other'.