This is a failed attempt to reproduce the "overload bit or bug"
Trying to understand why the overload of bitwise or in bitOpOnClass.chpl
causes a failure, I introduced a new function 'foobar', and similar
overloads of it as are defined for '|'; used it within one of the param
contexts where using '|' was causing problems; and provided a user
overload of it on a class to see if I could reproduce the issue, but
could not, so the issue must be deeper w.r.t. how '|' is used within the
internal modules...  Tucking this away for future reference.