Baby step forward
This changes from storing a default rectangular domain for the
upwards-facing domain in a reindexViewDom to storing the array
that was passed in, requiring it to be privatizable.  This
moved me an inch forward, but then the problem becomes the
setIndices() call which is correctly noted as being potentially
problematic in its comments.