Add int/uint ops for all widths
Thinking about this a bit more after the initial commit, it seems
as though one would want these semantics not just for the top of
the int/uint bitwidth hierarchy, but for the smaller sizes as well.
In fact, it raises a question as to whether the smaller sizes
should be errors by default rather than promoting to the larger int
sizes.