This is my renewed attempt to get rid of inner multiplies
Here, I use a param in DefaultRectangular to squash the inner multiply
by default until such time that we might conservatively need it.  I
know that this breaks due to the use of the => operator in BlockDist
slice, rankchange, and reindex operations, but wanted to remind myself
which tests did fail.