Make ascii() take an offset
The pattern ascii(mystring[i]) is really slow and inefficient in our
current implementation.  This commit adds a second optional argument
to ascii() to indicate an offset within the string argument, defaulting
to 0 for backwards compatibility.  Thus, the above can be written:
ascii(mystring, i) for faster exection.

I was thinking this would help the revcomp shootout() benchmark, but
made the classic error.  Since the calls like this aren't in the
inner loop, they're basically washed out by the overall execution
time.