Attempt to enable multiple bad 'param' initialization errors at once
Previously, when we hit a case where a 'param' was initialized with
a non-param argument, we generated one error and exited.  Here, I
attempt to generate multiple errors by switching an INT_FATAL() to
an INT_FATAL_CONT() in combination with marking the first param
as being a non-param (since, effectively, it isn't).  Failure to
de-mark it like this resulted in other errors that were more
confusing because we'd end up calling 'param' versions of functions
with things that the compiler couldn't process.  While this removal
of the flag is a little destructive in nature (throwing away
information), since we're about to error out anyway, it seems worth
it for the sake of giving users more error messages (which is a
common complaint against the compiler).

TODO:
[] testing