This tried to simplify things by only resolving on the generic branch
It doesn't fix key futures like whereClauseCantFold, though, because
those cases were handled by the where resolution on the concrete
branch after the generic had been instantiated.  The quiet made sure
we didn't trigger an error on the generic branch and then they
fell out of the candidate list once they were concretized.  So
this is perhaps a false branch.

New idea: Can we treat functions on the concrete branch differently
if they're instantiated or not?