Further optimizations of blk(rank) == 1
In folding out innerMult optimizations, I started wondering what it
would take to simply stop storing the blk(rank) element given that
it's now always 1.  This patch essentially does that (_almost_) by
declaring 'blk' to be '(rank-1)*int' and eliminating any resulting
accesses to blk(rank), either by optimizing them away or replacing
them with '1'.

Unfortunately, the change was not quite as clean as I hoped because:

(a) we don't support 0-tuples at present, so I had to special-case the
    declaration of 'blk' for rank==1 arrays.  For now, I wrote this as
    a 'real' because it made it simple to find cases that were still
    referring to blk(rank) for the 1D case.  Ideally this would be
    'void'.  This shouldn't be merged to master until it's at least
    made a less arbitrary type like 'int' or 'bool'

(b) several of the distributions that do RAD optimizations sadly
    duplicate the type of the 'blk' field informally by declaring a
    SENTINEL variable of the same type, yet by replicating the type
    expression.  This required me to special-case each of those
    delarations as well.  This could be improved somewhat if we had a
    little helper 'type' function which returned the type of the 'blk'
    field and used this both to declare the original fields and the
    SENTINEL values.

The bulk optimization also was a little messier to update to the new
optimized form than the other cases because it's one of the few cases
to use a non-param indices into the 'blk' accesses, making it harder
to peel off blk(rank) accesses via code transformations.  This left
things a little uglier than ideal/other cases.