Make alignment of non-stridable ranges zero
Prior to this commit, alignment on non-stridable ranges was simply
the low bound.  This commit changes it to 0 on the argument that
we're storing a value in 0..#1 which seems "better".  But, for
1..n by 1 we store '1' as the alignment currently.  So it seems
like we should either keep storing 'lo' as the alignment for
consistency or to store all alignments as 0..#stride (which I
wasn't quite ready to take up yet).