Convert assertion to early return
In my first draft, I was getting new assertion failures in this
routine, so in my "can I fix this in five minutes" mindset, I
changed it to an early return instead, thinking "we're now
considering routines that are technically dead code, so this
might make sense."  This resulted in 10% of the tests failing
with internal errors and about 20 getting compiler timeouts.

Looks like this'll take more than five minutes (though I'm still
not convinced it's more than a day's work).