Snapshot of attempt to scan iterator records by computing result shape first
This was a first stab at trying to support scans of promoted array
expressions like (A: int) by using Vass's shape introspection routines
on iterator records to declare the array shape early and then passing
it into the scan routines.  What I failed to anticipate is that we
don't support things like random access indexing on iteratorRecords
(currently), so I can't just reuse the existing code as I'd hoped.

I think the most likely way forward here would be to make
iteratorRecords resulting from array (and domain?) promotions work
more like a pseudo-array (an array view) that supports anything you
can do with an array by just forwarding to the underlying array(s) and
applying operations on the way out.  E.g., dsiAccess for such an array
would call A.dsiAccess() and then cast the result to int.

The other way to go that I was thinking about was to create a follower
iterator that knew how to apply a scan to a range of values and
rewrite the scan code to use a zippered iteration over the result
array and the input data set.  Since iteratorRecords support
leader/follower iterators (right?), this could take care of all of the
computation of tasks, chunks, etc. cleanly rather than explicitly as
I'm doing today.  The stumbling block here is that each task would
need to return the final value of its chunk up to the top level to
support the scanning of all the tasks' results, but it doesn't seem
like we could get this data out of a forall in a clean / general way
(at least that I've been able to figure out so far).  Though it also
seems like an attractive pattern to try and support, just tricky...