Change FCFs' type
Prior to this change, a declaration like:

var x = foo;

for a function like:

proc foo(int): int { ... }

would essentially be like:

var x: func(int, int) = foo;

That is, 'x' would have a general type (a function that takes an int
and returns an int) and a specific value (foo).  With this change,
it essentially makes x into:

var x: foo_type = foo;

where foo_type is a subclass of 'func(int, int)' that uniquely
describes the function foo.  In essence, this makes the fact that each
function has a unique type visible to the user.  As a result, several
tests have to change in a way that makes them less productive, and
there are a few I haven't gotten working quite yet.