Remove CHPL_AUX_FILESYS
This was a start at an attempt to remove the deprecated CHPL_AUX_FILESYS
before realizing that it seems we still rely on it for Lustre.  Best
guess at present is that we want to continue to support Lustre, but not
via this variable, which would suggest introducing some other variable
to set it, like CHPL_LUSTRE or the like...  That's bigger than I can
take on today for the 1.30 feature freeze, though.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>