Make (standalone) forall loops over arrays of locales parallel
This PR makes 'forall loc in myArrayOfLocales' into a coforall+on
style loop over the array of locales.  I put this together as an
alternative to a concept we've had in the past of having a special
distribution that would be used for arrays of locales to get this
sort of behavior.  In this PR, I'm wondering, "What if this were
simply the default behavior for arrays of locales... would that
be problematic?  Do we ever rely on such forall loops today?"

This just updates the standalone case; additional work would be
required to get leaders doing the same thing.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>